taste_backend/requirements.txt
asgiref==3.6.0
Babel==2.14.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.1
dj-rest-auth==2.2.7
Django==4.1.5
django-cors-headers==4.3.1
django-money==3.4.1
django-rest-framework==0.1.0
django-stub==0.1
django-stubs==5.0.2
django-stubs-ext==5.0.2
django-types==0.19.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Faker==22.6.0
firebase-admin==6.4.0
google-api-core==2.15.0
google-api-python-client==2.115.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==22.0.0
httplib2==0.22.0
idna==3.6
msgpack==1.0.7
neovim==0.3.1
packaging==24.1
proto-plus==1.23.0
protobuf==4.25.2
psycopg2-binary==2.9.9
py-moneyed==3.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.6.0
pynvim==0.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2022.7
requests==2.31.0
rsa==4.9
setuptools==70.0.0
six==1.16.0
sqlparse==0.4.3
types-psycopg2==2.9.21.20240417
types-PyYAML==6.0.12.20240311
typing_extensions==4.12.1
uritemplate==4.1.1
urllib3==2.1.0
